backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.2
python-multipart==0.0.20
boto3==1.36.10
google-genai==1.5.0
elevenlabs==1.50.5
mem0ai==0.1.43
pinecone-client==5.1.0
open-clip-torch==2.29.0
Pillow==11.1.0
numpy==2.2.3
ddtrace==2.20.0
datadog==0.50.2
pydantic==2.10.5
python-dotenv==1.0.1
httpx==0.28.1
